+++ components-starters/components-starter-object-storage/src/main/java/com/skrstop/framework/components/starter/objectStorage/entiry/CosStorageTemplateSign.java
@@ -1,0 +1,17 @@
+package com.skrstop.framework.components.starter.objectStorage.entiry;
+
+import lombok.Getter;
+import lombok.Setter;
+
+/**
+ * @author 蒋时华
+ * @date 2024-03-18 18:39:27
+ * @since 1.0.0
+ */
+@Getter
+@Setter
+public class CosStorageTemplateSign extends StorageTemplateSign {
+    private String tmpSecretId;
+    private String tmpSecretKey;
+    private String sessionToken;
+}